MILESTONE 3/requirements.txt
openai>=1.0.0
langgraph>=0.0.1
python-dotenv>=1.0.0
requests>=2.31.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
faiss-cpu>=1.7.0
tiktoken>=0.5.0
streamlit>=1.24.0
